move ha const out of lib
